Changed to use the `Decimal` module for better

+++ src/pymathematic/constants.py
@@ -1,5 +1,6 @@
 import math
+from decimal import Decimal
 
 e: float = math.e
 pi: float = math.pi
-phi: float = (1 + 5 ** 0.5) / 2
+phi: float = float((1 + 5 ** Decimal(0.5)) / 2)
